requirements.txt
Flask==3.1.0
Flask-Limiter==3.8.0
Flask-WTF==1.2.2
Flask-Talisman==1.1.0
requests==2.32.3
python-dotenv==1.0.1
bleach==6.2.0
